improve jenkins workflow
